requirements.txt
pororo
python-mecab-ko
torchserve
torch-model-archiver
